requirements.txt
Flask>=2.0.1,<3.0.0
fer>=22.4.0,<23.0.0
opencv-python-headless>=4.5.5,<5.0.0
numpy>=1.20.0,<2.0.0
Pillow>=10.0.0,<11.0.0
tensorflow>=2.8.0,<3.0.0
matplotlib>=3.5.1,<4.0.0